apps/api/requirements.txt
fastapi
uvicorn
celery
redis
instaloader
yt-dlp
ffmpeg-python
groq
tavily-python
google-api-python-client
google-auth-oauthlib
supabase
pydantic
pydantic-settings
httpx
tenacity
python-dotenv
python-multipart
